Adding connect_options omit_default_port
Fog uses a gem called excon for making http requests. This gem defaults
to adding a port in host headers of the request with the url. This
doesn'tw work with platform1 on skyscape, and also is not the right way
of doing things. excon provides and option of omit_default_port which
removes the port from host headers. We can pass this to excon via fog's
connection_options