[#55141908] Force guest customisation
Ensure that guest customisation is enabled even if the vApp template and
subsequently provisioned VM has it disabled.

This will cause the API to throw a 400 response if it thinks that the
version of VMware Tools on the guest doesn't support customisation. I think
this is more desirable, because it makes it apparent that the VM won't have
the correct hostname or static IP address.