[#55141908] Cycle power to identify VMware Tools
There is some problem with our template which causes Guest Customization to
be disabled, both for the template and newly provisioned VMs. Which means
that network and hostname configurations are not applied.

Preceding commit 4e81b43 forces the flag to be turned back on after a
vApp/VM has been provisioned. But it fails with a 400 response because
vCloud doesn't think that the VM has a suitable version (or any install, at
all) of VMware Tools installed.

However we have now produced a template with an OSP version of VMware Tools
that works after a VM has been booted once, then shutdown and the flag
changed. So take advantage of this by doing the same steps during
provisioning.

After booting the VM, and before applying customisations, it waits 90s for
vCloud to populate the `VMWareTools` attribute. When this is present vCloud
has recognised that VMware Tools is present. The VM is then powered off
(caveat in comments) and customisation proceeds.

Tested against SS Platform1 with alphagov/boxgrinder-appliances@60550e5