[#55141908] Replace #power_off with #undeploy
The `#power_off` method leaves the VM in a "Partially Powered Off" state
which causes the Guest Customizations to not be applied. If you look at the
VM properties it will say something to the effect of "Customisations are out
of sync with the guest OS. They will be synchronized the next time the VM is
powered on" (paraphrased)

It seems that `#undeploy` is preferred for correctly shutting down a VM. The
network/hostname settings are then applied without manual intervention. I
believe this will also default to a graceful shutdown instead of pulling the
power.

Despite it's scary name, the documentation seems to confirm that it doesn't
actually delete anything:

http://pubs.vmware.com/vcd-51/index.jsp#operations/POST-UndeployVApp.html