Explain why this project is deprecated
While writing the vCloud Tools blog post I realised there was a lot of detail about why we moved on from previous iterations that would actually not be that relevant to the majority of readers. I think it might be better to put that information here, so that if people are interested enough to go and have a look, it is explained here.

Also offers information for people who find this repo not via the blog post.